requirements.txt
annotated-types==0.6.0
customtkinter @ git+https://github.com/CaioEmPessoa/CustomTkinter.git@0f703b28e647ad6d9dddb3e9f2ad794a7a7db0a7
darkdetect==0.8.0
packaging==23.2
Pillow==10.1.0
typing_extensions==4.8.0
icoextract==0.1.4